Refactored to use new Product extension method instead of aggregate

+++ Extensions.cs
@@ -47,9 +47,9 @@ namespace ProjectEulerSharp
             return true;
         }
 
-        public static int Product(this IEnumerable<int> collection)
+        public static long Product(this IEnumerable<int> collection)
         {
-            return collection.Aggregate(1, (p, v) => p * v);
+            return collection.Aggregate(1L, (p, v) => p * v);
         }
 
         public static long Product(this IEnumerable<long> collection)
